Move useRedirectToHistoryPageOnInvalidTransaction to own file
