+++ client/my-sites/purchases/billing-history/use-redirect-to-history-page-on-invalid-transaction.tsx
@@ -1,0 +1,34 @@
+/**
+ * External dependencies
+ */
+import { useEffect, useRef } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+
+/**
+ * Internal dependencies
+ */
+import { getBillingHistoryUrlFor } from '../paths';
+import isPastBillingTransactionError from 'state/selectors/is-past-billing-transaction-error';
+import { clearBillingTransactionError } from 'state/billing-transactions/individual-transactions/actions';
+
+export default function useRedirectToHistoryPageOnInvalidTransaction(
+	siteSlug: string,
+	receiptId: number
+) {
+	const transactionFetchError = useSelector( ( state ) =>
+		isPastBillingTransactionError( state, receiptId )
+	);
+	const reduxDispatch = useDispatch();
+	const didRedirect = useRef( false );
+
+	useEffect( () => {
+		if ( didRedirect.current ) {
+			return;
+		}
+		if ( transactionFetchError ) {
+			didRedirect.current = true;
+			reduxDispatch( clearBillingTransactionError( receiptId ) );
+			page.redirect( getBillingHistoryUrlFor( siteSlug ) );
+		}
+	}, [ transactionFetchError, receiptId, reduxDispatch, siteSlug ] );
+}
